chore: moving path back
