+++ v2/assets/metric-state/service-monitor.yaml
