+++ v2/assets/metric-state/service.yaml
