+++ v2/assets/prometheus-operator/operator-certs-ca-bundle.yaml
