+++ v2/assets/prometheus-operator/service.yaml
