+++ v2/assets/prometheus/additional-scrape-configs.yaml
