+++ v2/assets/prometheus/htpasswd-secret.yaml
