+++ v2/assets/prometheus/kube-rbac-proxy-secret.yaml
