+++ v2/assets/prometheus/kube-state-service-monitor.yaml
