+++ v2/assets/prometheus/kubelet-serving-ca-bundle.yaml
