+++ v2/assets/prometheus/prometheus-datasources-secret.yaml
