+++ v2/assets/prometheus/prometheus-rules.yaml
