+++ v2/assets/prometheus/proxy-secret.yaml
