+++ v2/assets/prometheus/service.yaml
