+++ v2/assets/prometheus/serving-certs-ca-bundle.yaml
