+++ v2/assets/razee/razee-namespace.yaml
