+++ v2/assets/razee/remote-resource-s3.yaml
